@@ -3,7 +3,10 @@ import 'package:hola/pages/home/home_page_checkbox_scaffold.dart';
 import 'package:hola/pages/home/home_page_counter_column.dart';
 
 class HomePage extends StatelessWidget {
-  HomePage({Key key, this.title}) : super(key: key);
+  HomePage({
+    Key? key,
+    required this.title,
+  }) : super(key: key);
   final String title;
 
   @override
